requirements.txt
annotated-types==0.6.0
APScheduler==3.10.4
aiohttp==3.9.5
anyio==4.2.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
Deprecated==1.2.14
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
importlib-resources==6.1.1
numpy==1.26.3
orjson==3.10.1
packaging==23.2
pandas==2.2.1
paramiko==2.11.0
pycparser==2.22
pydantic==2.6.3
pydantic_core==2.16.3
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.9.0
tzdata==2023.4
uvicorn==0.27.0
uvloop==0.19.0
wrapt==1.16.0
